Make WifiManage headers private
